python_frontend/requirements.txt
pywebview
flask
requests
PySide6
numpy
SpeechRecognition
pyaudio
